Started updateepisode on basecontentpage, Switching branches to look at HasJournal

+++ DABApp/Droid/DabSockets/droidWebSocket.cs
@@ -90,17 +90,17 @@ namespace DABApp.Droid.DabSockets
                 if (firstEpObject.favorite != action.favorite)
                 {
                     actionType = "favorite";
-                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, action.position, action.listen, action.favorite);
+                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, (int)action.position, action.listen, action.favorite);
                 }
                 else if (firstEpObject.listen != action.listen)
                 {
                     actionType = "listened";
-                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, action.position, action.listen, action.favorite);
+                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, (int)action.position, action.listen, action.favorite);
                 }
                 else if (firstEpObject.position != action.position)
                 {
                     actionType = "pause";
-                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, action.position, action.listen, action.favorite);
+                    await AuthenticationAPI.CreateNewActionLog(action.episodeId, actionType, (int)action.position, action.listen, action.favorite);
                 }
                 else
                 {
